requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.5
pymongo==4.10.1
python-dotenv==1.0.1
Flask-WTF==1.2.1
email-validator==2.1.1
